rpgmxp_tool::commands::compile_assets::file_sink

Enum FileSink

pub enum FileSink {
    Dir {
        base_path: PathBuf,
    },
    Rgssad {
        writer: Writer<BufWriter<File>>,
    },
}

An abstraction of a file sink over dir and rgssad output formats.

Variants

Dir

Fields

base_path: PathBuf

Rgssad

Fields

writer: Writer<BufWriter<File>>

Implementations

impl FileSink

pub fn new_dir(path: &Path, overwrite: bool) -> Result<Self>

Create a new file sink for a directory

pub fn new_rgssad(path: &Path, overwrite: bool) -> Result<Self>

Create a new file sink for an rgssad file

pub fn write_file<R>( &mut self, path_components: &[&str], size: u32, reader: R, ) -> Result<()>
where R: Read,

Write a file.

pub fn finish(&mut self) -> Result<()>

Finish and close this file sink.