rpgmxp_tool::commands::extract_assets::file_entry_iter

Enum FileEntryIter

pub enum FileEntryIter {
    WalkDir {
        input_path: PathBuf,
        iter: IntoIter,
        game_kind: GameKind,
    },
    Rgssad {
        reader: Reader<File>,
        game_kind: GameKind,
    },
}

A lending iter over files.

Variants

WalkDir

Fields

input_path: PathBuf

iter: IntoIter

game_kind: GameKind

Rgssad

Fields

reader: Reader<File>

game_kind: GameKind

Implementations

impl FileEntryIter

pub fn new<P>(path: P) -> Result<Self>
where P: AsRef<Path>,

Create a new iter from a path.

This will determine whether the path is a dir or an rgssad.

pub fn new_walkdir_path<P>(path: P) -> Result<Self>
where P: AsRef<Path>,

Create a new iter from the given dir path.

pub fn new_rgssad_path<P>(path: P) -> Result<Self>
where P: AsRef<Path>,

Create a new iter from the given rgssad path.

pub fn new_rgssad_file(file: File, game_kind: GameKind) -> Result<Self>

Create a new iter from the given rgssad file.

pub fn next_file_entry(&mut self) -> Result<Option<FileEntry<'_>>>

Get the next file entry.

pub fn game_kind(&self) -> GameKind

Get the determined game kind